+++ ios/QliroOneViewManager.h
@@ -1,0 +1,16 @@
+//
+//  QliroOneViewManager.h
+//  qliroone_reactnative
+//
+//  Created by Elliot Rask on 2022-07-03.
+//
+
+#import "RCTViewManager.h"
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface QliroOneViewManager : RCTViewManager
+
+@end
+
+NS_ASSUME_NONNULL_END