+++ src/QliroOneNativeComponent.tsx
@@ -1,0 +1,5 @@
+import { requireNativeComponent } from 'react-native';
+import { NativeCheckoutIOS } from './QliroOneTypes';
+
+export const QlirOneNativeCheckout: typeof NativeCheckoutIOS =
+  requireNativeComponent('QliroOneCheckout');