+++ src/QliroOneProps.ts
@@ -1,4 +1,4 @@
-import { Order, PurchaseRedirectOptions } from './models';
+import { Module } from "./models";
 
 export interface QliroOneProps {
   /**
@@ -7,32 +7,26 @@ export interface QliroOneProps {
   orderHtml: string | undefined;
 
   /**
-   * baseUrl is used in order to identify the main frame's origin in a trustworthy way,
-   * you need to provide a valid HTTP or HTTPS base URL to set the origin.
+   * Set to true to enable scroll, otherwise the checkout will size itself after its content.
+   * This should not be set to true if you are nesting the checkout inside of another scrollview.
    */
-  baseUrl?: string;
+  isScrollEnabled?: boolean;
 
   /**
-   * Called after the updateOrders action has been called when Qliro One has synced its orders.
-   * This might be called multiple times and should return true when
-   * the Qliro One and the app is in sync. Returning true will unlock the Checkout.
-   * @param order - The order
-   * @param rest - Possible extra arguments not yet supported by SDK types. [Read more]{@link https://developers.qliro.com/docs/qliro-one/checkout-features/update-order}
-   * @returns {boolean} - true if the order is successfully synced otherwise false
-   **/
-  onOrderUpdated?: (order: Order, ...rest: any[]) => boolean;
-
-  /**
-   * Called when Qliro is about to show the success page after a successful payment.
-   * If you set this function you will override the default behaviour of redirecting to your specified success url.
-   * @param options - An object with data to be used to customize the redirect to your liking.
-   **/
-  onCompletePurchaseRedirect?: (options: PurchaseRedirectOptions) => void;
-
-  /**
-   * Sets to enable scroll, otherwise the height will be as tall as required.
+   * Lock the Qliro One frontend and disable user interaction.
    */
-  scrollEnabled?: boolean;
+  locked?: boolean;
+
+  /**
+   * Used to exclude optional modules from displaying on the result page (thank you page).
+   */
+  excludedResultModules?: Module[];
+
+  /**
+   * When enabled, Qliro One will scroll to the selected payment or shipping option after a customer have chosen one in the expanded list. (default: false)
+   * This have no effect when isScrollEnabled is true.
+   */
+  isCheckoutScrollEnabled?: boolean;
 
   /**
    * Called when something is logged